validation/pyproject.toml
[project]
name = "rag-tutor"
version = "0.1.0"
description = "RAG eval pipeline with DeepEval/BLEURT and FAISS"
requires-python = "==3.12.*"
dependencies = [
    "streamlit>=1.45.1",
    "langchain-community==0.3.24",
    "langchain-openai==0.3.18",
    "langchain-core==0.3.61",
    "faiss-cpu==1.11.1",
    "FlagEmbedding==1.3.4",
    "transformers==4.51.1",
    "torch==2.4.1",
    "python-dotenv==1.1.0",
    "PyYAML==6.0.2",
    "requests==2.32.3",
    "einops==0.8.1",
    "rank-bm25==0.2.2",
    "pandas==2.2.3",
    "numpy==1.26.4",
    "scikit-learn==1.6.1",
    "tqdm==4.67.1",
    "deepeval==3.0.8",
    "bleurt-pytorch==0.0.1",
    "markdownify==1.1.0",
    "sentence-transformers==5.1.2",
]

[project.optional-dependencies]
gpu = [
    "vllm>=0.8.5",
    "faiss-gpu-cu12>=1.11.0",
]
dev = [
    "ruff>=0.11.13",
    "pre-commit>=4.0.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100
extend-exclude = ["faiss_data", "research/data"]

[tool.ruff.lint]
select = ["E","W","F","I","B","C4","UP"]
ignore = ["E501","B008","C901"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
